% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdi.R
\name{agrep_across_all_field_terms}
\alias{agrep_across_all_field_terms}
\title{Search for an ontology term across all the possible GRDI fields.}
\usage{
agrep_across_all_field_terms(x, ...)
}
\arguments{
\item{x}{query to search the ontology terms with.}

\item{...}{passed onto \code{\link[=agrep]{agrep()}} for fuzzy matching.}
}
\description{
In case you want to search for a specific ontology term across all the
GRDI fields. Will print out the matches and the field to which that
ontology term belongs.
}
