% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{compare_lookup_tables}
\alias{compare_lookup_tables}
\title{Check for discrepancies in the various forms of the GRDI template}
\usage{
compare_lookup_tables()
}
\value{
A dataframe of discrepancies.
}
\description{
There are 3 sources of information for the GRDI ontology: the Master
Template, the Excel Spreadsheet template, and the yaml for the
Data Harmonizer. Use this function to check for discrepancies between
the three.
}
\details{
The function does it all, including data import.
}
