% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{compare_one_lookup_field}
\alias{compare_one_lookup_field}
\title{Internal: compare a single field and return decrepancies.}
\usage{
compare_one_lookup_field(df, field)
}
\description{
Used in \link{compare_lookup_tables}
}
