% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr.R
\name{convert_GRDI_ont_to_vmr_ids}
\alias{convert_GRDI_ont_to_vmr_ids}
\title{Convert GRDI ontology term into VMR ontology indexes}
\usage{
convert_GRDI_ont_to_vmr_ids(db, x)
}
\arguments{
\item{db}{\link{DBI} connection to the VMR}

\item{x}{Vector of ontology terms in the form of "term name \link{ONT:00000000}"}
}
\description{
Ontology terms in the VMR are indexed in its ontology_terms
table. To insert these terms into their respective columns, they
must be converted into these index ids. This function handles this
from within R by querying the ontology_term table.
}
