% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_parsing.R
\name{excel_lookup_values}
\alias{excel_lookup_values}
\title{Get ontology values from the excel sheet.}
\usage{
excel_lookup_values(template_file)
}
\description{
Get ontology values from the excel sheet.
}
