% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdi.R
\name{extract_ontology_name}
\alias{extract_ontology_name}
\title{Get ontology name from an ontology ID}
\usage{
extract_ontology_name(x)
}
\arguments{
\item{x}{vector of ontology IDs}
}
\description{
Given a vector of ontology IDs (In the form of ONTOLOGY:00000000),
returns the lowercase name of the ontology ("ontology")
}
