% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdi.R
\name{get_all_field_ontology_terms}
\alias{get_all_field_ontology_terms}
\title{Get a list of all the ontology terms used in the GRDI standard.}
\usage{
get_all_field_ontology_terms()
}
\description{
Get a list of all the ontology terms used in the GRDI standard.
}
