% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdi.R
\name{get_fields_of_group}
\alias{get_fields_of_group}
\title{Return the fields of a grdi categery}
\usage{
get_fields_of_group(x)
}
\arguments{
\item{x}{The GRDI group. Can be an approximate match}
}
\description{
Use this to return the fields (and their associated data) that belong
to one of the GRDI field categories.
}
