% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdi.R
\name{get_fields_with_menus}
\alias{get_fields_with_menus}
\title{Get a vector of GRDI fields that have pick-lists}
\usage{
get_fields_with_menus()
}
\description{
To get a list of the GRDI fields that have mandatory values (ontologies)
to be selected
}
