% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdi.R
\name{get_info}
\alias{get_info}
\title{Return the description and comments tags of a GRDI field}
\usage{
get_info(field)
}
\arguments{
\item{field}{GRDI field to query.}
}
\description{
Return the description and comments tags of a GRDI field
}
