% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdi.R
\name{get_null_value}
\alias{get_null_value}
\title{Get a GRDI ontology null value}
\usage{
get_null_value(x = NULL)
}
\arguments{
\item{x}{query to search for a specific ontology term}
}
\description{
Returns a valid NULL value. If no query is supplied to search for a specific
ontology term, than a list of all the possible NULL values are supplied.
}
