% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{get_terms_from_excel_and_yaml_sources}
\alias{get_terms_from_excel_and_yaml_sources}
\title{Return a df of ontology terms and fields from all three sources}
\usage{
get_terms_from_excel_and_yaml_sources()
}
\description{
Get a datagrame with the ontology terms, ids, and fields of the
GRDI standard, from the Master excel sheet, the DataHarmonizer YAML,
and the Excel template
}
