% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr.R
\name{replace_null_params_with_na}
\alias{replace_null_params_with_na}
\title{If a list of parameters for \link[DBI:dbBind]{DBI::dbBind} contains NULL, set to NA}
\usage{
replace_null_params_with_na(params)
}
\arguments{
\item{params}{List of vectors to pass onto \link[DBI:dbBind]{DBI::dbBind}}
}
\description{
For \link[DBI:dbBind]{DBI::dbBind}, the parameters argument takes a list of the values
to bind to the placeholders. The vectors comprising this list must all
be the same length. Any argument that has not been initialized must be set
to a vector of NAs, of same length as the other vectors.
}
