% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdi.R
\name{replace_with_GRDI_term}
\alias{replace_with_GRDI_term}
\title{Replace values in a column with GRDI ontologies}
\usage{
replace_with_GRDI_term(
  df,
  col_name,
  term_query,
  data_query = NULL,
  term_query_dist = 0,
  data_query_dist = 0,
  select_GRDI_term = NULL
)
}
\arguments{
\item{df}{Dataframe}

\item{col_name}{column name of the dataframe to replace over}

\item{term_query}{For looking up the ontology term we want to replace with.}

\item{data_query}{The value in the data that we want replace}

\item{term_query_dist}{Distance passed onto \code{\link[=agrep]{agrep()}}, for fuzzy matching
the desired ontology term}

\item{select_GRDI_term}{If multiple results for an ontology term are returned,
you may select one manually.}

\item{data_query_data}{Distance passed onto \code{\link[=agrep]{agrep()}}, for fuzzy matching
the value in the data.}
}
\description{
Given a column in a dataframe, this function will find the supplied term
in the data and replace it with an appropriate ontology from the GRDI
standard. Supports fuzzy matching via \code{\link[=agrep]{agrep()}}
}
