% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr.R
\name{select_and_rename_to_vmr}
\alias{select_and_rename_to_vmr}
\title{Renaming a GRDI formatted dataframe to a VMR table}
\usage{
select_and_rename_to_vmr(db, df, vmr_table)
}
\arguments{
\item{db}{A DBI connection to the VMR}

\item{df}{Dataframe, formatted according to the GRDI spec}

\item{vmr_table}{The table in the VMR database to convert into}
}
\value{
A dataframe with columns that match the fields of the selected
table in the VMR
}
\description{
Take a dataframe imported from a GRDI specification template and
convert it into a VMR table ready for insertion into the database.
This function depends on an updated mapping YAML.
}
