% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr.R
\name{sendBindFetch}
\alias{sendBindFetch}
\title{Wrapper for \link[DBI:dbSendQuery]{DBI::dbSendQuery}, \link[DBI:dbBind]{DBI::dbBind}, and \link[DBI:dbFetch]{DBI::dbFetch}}
\usage{
sendBindFetch(db, sql, params, verbose = FALSE)
}
\arguments{
\item{db}{A \link{DBI} connection}

\item{sql}{An sql statement, as a string. Use "$n" as placeholders for the
parameters}

\item{params}{A list of the corresponding parameters.}
}
\value{
A dataframe (tibble) of the results of the query
}
\description{
An operation sufficiently common I wrapped it into a function
}
