% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdi.R
\name{separate_ontology_terms}
\alias{separate_ontology_terms}
\title{Separate GRDI ontology terms into two columsn with the Term and ID}
\usage{
separate_ontology_terms(data, col)
}
\arguments{
\item{data}{a data frame}

\item{<tidy-select>}{Column to separate}
}
\description{
A tidyverse eval function. Uses \link[tidyr:separate_wider_delim]{tidyr::separate_wider_regex} to
separate a column of form "Term Name [ONTOLOGY:00000000]" in the
Term name in one column and the ID into another
}
