% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{test_for_term_descrepancies}
\alias{test_for_term_descrepancies}
\title{Test for terms that are duplicated or misnamed}
\usage{
test_for_term_descrepancies()
}
\value{
A dataframe of ontology terms and the tables they are found in
}
\description{
Test for duplicated or misnamed ontology terms between the 3 forms of the
GRDI template: the Excel Template , the YAML, and the Master Reference Guide.
}
