% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr.R
\name{vmr_map_as_dataframe}
\alias{vmr_map_as_dataframe}
\title{Maps GRDI cols to their respective Tables and Fields in the VRM}
\usage{
vmr_map_as_dataframe()
}
\description{
The VMR adheres as closely as possible to the naming conventions of the
GRDI data specification. This function takes a YAML file that maps the
GRDI columns to the VMR's respective representation of these fields
in their table and column. This YAML must be maintained separately,
and is found in \verb{data-raw/} of this package's source.
}
